requirements.txt
requests
streamlit
numpy
Pillow
flask
ultralytics
pandas
streamlit-authenticator